Changed DESIRE_ASIZEOF() macro to return an uint32_t instead of size_t

+++ DESIRE-Engine/src/Engine/Common.h
@@ -42,15 +42,15 @@
 #define DESIRE_UNUSED(X)					((void)X)
 
 // Array sizeof which returns the number of elements in a static array
-template<typename T, size_t N>
-constexpr size_t DesireArraySizeHelper(T(&)[N])
+template<typename T, uint32_t N>
+constexpr uint32_t DesireArraySizeHelper(T(&)[N])
 {
 	return N;
 }
 #define DESIRE_ASIZEOF(X)					DesireArraySizeHelper(X)
 
 // Compile-time check to make sure the array has the same number of elements as 'count'
-#define DESIRE_CHECK_ARRAY_SIZE(X, COUNT)	static_assert(DESIRE_ASIZEOF(X) == (size_t)COUNT, "Array size doesn't match the number of elements");
+#define DESIRE_CHECK_ARRAY_SIZE(X, COUNT)	static_assert(DESIRE_ASIZEOF(X) == static_cast<uint32_t>(COUNT), "Array size doesn't match the number of elements");
 
 // Prevent copy and move
 #define DESIRE_NO_COPY_AND_MOVE(Class)			\
